+++ explorer/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - Add Tokens page [#3415]
+- Add Copy button next to address (Account page) [#3698]
 - Add amount value to Transaction Details [#3699]
 
 ### Changed
@@ -245,6 +246,7 @@ and this project adheres to
 [#3676]: https://github.com/dusk-network/rusk/issues/3676
 [#3692]: https://github.com/dusk-network/rusk/issues/3692
 [#3694]: https://github.com/dusk-network/rusk/issues/3694
+[#3698]: https://github.com/dusk-network/rusk/issues/3698
 [#3699]: https://github.com/dusk-network/rusk/issues/3699
 [#3720]: https://github.com/dusk-network/rusk/issues/3720